[refactor]: extract the shared in-memory query pipeline;
The filter/sort/paginate sequence was written out separately in the
static strategy, the localStorage strategy and the README examples, with
slightly different edge-case behaviour in each copy. applyQuery in
lib/utils/query.ts is now the single implementation; the static strategy
uses it as of this commit.

Behavioural fix picked up from the stricter copy: rows whose value is
missing no longer match a filter via the literal string 'undefined', and
missing values sort last in either direction.

Refs #4
